projetos/logistica_transporte/5_analise_geoespacial_de_areas_de_origem_e_destino_visualizacao_de_dados/src/requirements.txt
# Frameworks e Bibliotecas
# Python 3.8
#accelerate>=0.26.0,<0.27.0
#arch
#boto3==1.18.0
#catboost>=1.2.7,<2.0
#celery==5.3.5
#chardet==5.1.0
#click>=8.0.1,<9.0
#cmake==3.25.0
cmdstanpy==1.0.4
#cython>=0.29.21,<0.30
#dask[dataframe]>=2021.9.1
#dask-ml>=2022.1.1,<2023.0
#dlib==19.24.6
#emoji>=2.6.0
#face-recognition
#fastapi
fastparquet==0.5.0
#fasttext>=0.9.2,<1.0
#fer
#flask>=2.0,<3.0
#flask-Caching==2.3.0
#flask_login==0.5.0
#flask_prometheus_metrics
#flasgger==0.9.5
#gensim>=4.0.1,<5.0
geopandas==0.13.2
#geopy
holidays==0.14.2
#h5py==3.8.0
#imbalanced-learn>=0.8.0,<0.10
#ipython==7.34.0
#joblib>=1.0.1,<2.0
#kafka-python==2.0.2
#keras>=2.8.0,<2.9.0
#langdetect
#lime==0.2.0.1
#lime-python==2.0.13
#lightgbm>=3.3.1,<4.0
matplotlib>=3.4.3,<3.5
#markupsafe>=2.1.1,<3.0
#mlflow==2.6.0
#moviepy==1.0.3
#nets==0.0.3.1
#nltk>=3.6.3,<4.0
numpy==1.24.4
#opencv-python==4.6.0.66
#opencv-contrib-python
#optuna==4.2.0
pandas>=1.3.3,<1.6.0
pandas-profiling>=3.1.0,<4.0
#pillow>=10.2.0,<10.3.0
#polars==0.20.16
#plotly==6.0.0
prophet==1.1.1
#prometheus_client
#prometheus-flask-exporter
pyarrow==4.0.1
pystan==2.19.1.1
#pyspark==3.1.2
#psycopg2==2.8.4
#psycopg2-binary==2.8.4
#PyQt5==5.15.4
#pytesseract
pyYAML==5.4.1
#pytest>=6.2.4,<7.0
#redis==5.1
#requests==2.32.3
#reportlab==3.6.13
scikit-learn>=1.0,<1.2.0
scikit-optimize>=0.10.2,<0.11.0
#scikeras>=0.10.0,<0.11.0
#scipy==1.10.0
seaborn>=0.11.2,<0.12
#setuptools>=49.6.0
#shap
#six>=1.15.0
#smart-open>=6.0.0
#spacy>=3.0.0,<4.0
#statsmodels==0.13.5
#streamlit
#symspellpy>=6.7.8
#tensorflow>=2.8.0,<2.9.0
#tensorflow-estimator>=2.8.0,<2.9.0
#textblob>=0.15.3,<0.16
#torch>=2.2.0,<2.3.0
#torchvision>=0.17.0,<0.18.0
#tqdm==4.59.0
#transformers>=4.21.0,<4.22.0
#tslearn==0.6.1
#typing-extensions>=3.7.4,<4.0.0
#ultralytics==8.0.0
#uvicorn
#vaderSentiment>=3.0.0
#werkzeug==2.0.3
#wordcloud>=1.8.1,<2.0
#wrapt>=1.11.0,<1.15
#yfinance==0.2.50
#xgboost==1.5.2

